bilingual_bridge/requirements.txt
# FastAPI + WS server
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0


# Audio + DSP
numpy==1.26.4
scipy==1.11.4


# Cloud SDKs / HTTP
httpx==0.27.0
python-dotenv==1.0.1


# Google Cloud (ASR + Translate)
google-cloud-speech==2.26.0
google-cloud-translate==3.15.4
#google-api-core==2.19.1

# Deepgram
#deepgram-sdk==3.2.8